move base Track object to new module
