+++ mirdata/ikala.py
@@ -21,6 +21,7 @@ import librosa
 import logging
 import numpy as np
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -63,7 +64,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData('ikala_index.json', _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """ikala Track class
 
     Args:
