+++ mirdata/medley_solos_db.py
@@ -31,6 +31,7 @@ import librosa
 import logging
 import os
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -80,7 +81,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData("medley_solos_db_index.json", _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """medley_solos_db Track class
 
     Args:
