+++ mirdata/orchset.py
@@ -20,6 +20,7 @@ import numpy as np
 import os
 import shutil
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -98,7 +99,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData('orchset_index.json', _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """orchset Track class
 
     Args:
