+++ mirdata/rwc_jazz.py
@@ -35,6 +35,7 @@ import csv
 import logging
 import os
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -114,7 +115,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData('rwc_jazz_index.json', _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """rwc_jazz Track class
 
     Args:
