+++ mirdata/rwc_popular.py
@@ -14,6 +14,7 @@ import logging
 import numpy as np
 import os
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -107,7 +108,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData('rwc_popular_index.json', _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """rwc_popular Track class
 
     Args:
