+++ mirdata/salami.py
@@ -17,6 +17,7 @@ import logging
 import numpy as np
 import os
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -78,7 +79,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData('salami_index.json', _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """salami Track class
 
     Args:
