+++ mirdata/tinysol.py
@@ -50,6 +50,7 @@ import librosa
 import logging
 import os
 
+import mirdata.track as track
 import mirdata.utils as utils
 import mirdata.download_utils as download_utils
 import mirdata.jams_utils as jams_utils
@@ -108,7 +109,7 @@ def _load_metadata(data_home):
 DATA = utils.LargeData("tinysol_index.json", _load_metadata)
 
 
-class Track(utils.Track):
+class Track(track.Track):
     """tinysol Track class
 
     Args:
