+++ mirdata/track.py
@@ -1,0 +1,45 @@
+# -*- coding: utf-8 -*-
+"""track object utility functions
+"""
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+import types
+
+MAX_STR_LEN = 100
+
+
+class Track(object):
+    def __repr__(self):
+        properties = [v for v in dir(self.__class__) if not v.startswith('_')]
+        attributes = [
+            v for v in dir(self) if not v.startswith('_') and v not in properties
+        ]
+
+        repr_str = "Track(\n"
+
+        for attr in attributes:
+            val = getattr(self, attr)
+            if isinstance(val, str):
+                if len(val) > MAX_STR_LEN:
+                    val = '...{}'.format(val[-MAX_STR_LEN:])
+                val = '"{}"'.format(val)
+            repr_str += "  {}={},\n".format(attr, val)
+
+        for prop in properties:
+            val = getattr(self.__class__, prop)
+            if isinstance(val, types.FunctionType):
+                continue
+
+            if val.__doc__ is None:
+                raise ValueError("{} has no documentation".format(prop))
+
+            val_type_str = val.__doc__.split(':')[0]
+            repr_str += "  {}: {},\n".format(prop, val_type_str)
+
+        repr_str += ")"
+        return repr_str
+
+    def to_jams(self):
+        raise NotImplementedError
