+++ mirdata/utils.py
@@ -29,43 +29,9 @@ from collections import namedtuple
 import hashlib
 import os
 import json
-import types
 
 
 MIR_DATASETS_DIR = os.path.join(os.getenv('HOME', '/tmp'), 'mir_datasets')
-
-
-class Track(object):
-    def __repr__(self):
-        properties = [v for v in dir(self.__class__) if not v.startswith('_')]
-        attributes = [
-            v for v in dir(self) if not v.startswith('_') and v not in properties
-        ]
-
-        repr_str = "Track(\n"
-
-        for attr in attributes:
-            val = getattr(self, attr)
-            if isinstance(val, str):
-                val = '"{}"'.format(val)
-            repr_str += "  {}={},\n".format(attr, val)
-
-        for prop in properties:
-            val = getattr(self.__class__, prop)
-            if isinstance(val, types.FunctionType):
-                continue
-
-            if val.__doc__ is None:
-                raise ValueError("{} has no documentation".format(prop))
-
-            val_type_str = val.__doc__.split(':')[0]
-            repr_str += "  {}: {},\n".format(prop, val_type_str)
-
-        repr_str += ")"
-        return repr_str
-
-    def to_jams(self):
-        raise NotImplementedError
 
 
 def md5(file_path):
