+++ tests/test_loaders.py
@@ -9,7 +9,7 @@ import sys
 import pytest
 
 import mirdata
-import mirdata.utils as utils
+import mirdata.track as track
 from tests.test_utils import DEFAULT_DATA_HOME
 
 DATASETS = [importlib.import_module("mirdata.{}".format(d)) for d in mirdata.__all__]
@@ -66,7 +66,7 @@ def test_track():
         assert track_default._data_home == os.path.join(
             DEFAULT_DATA_HOME, dataset.DATASET_DIR)
 
-        assert isinstance(track_default, utils.Track)
+        assert isinstance(track_default, track.Track)
 
         assert hasattr(track_default, 'to_jams')
