+++ tests/test_utils.py
@@ -143,44 +143,3 @@ def test_validator(mocker, mock_check_index, missing_files, invalid_checksums):
 
 def test_get_default_dataset_path():
     assert '/tmp/mir_datasets/data_home' == utils.get_default_dataset_path('data_home')
-
-
-def test_track_repr():
-
-    class TestTrack(utils.Track):
-        def __init__(self):
-            self.a = 'asdf'
-            self.b = 1.2345678
-            self.c = {1: 'a', 'b': 2}
-            self._d = 'hidden'
-            self.e = None
-
-        @property
-        def f(self):
-            """ThisObjectType: I have a docstring"""
-            return None
-
-        @property
-        def g(self):
-            """I have an improper docstring"""
-            return None
-
-        def h(self):
-            return "I'm a function!"
-
-    expected1 = """Track(\n  a="asdf",\n  b=1.2345678,\n  """
-    expected2 = """c={1: 'a', 'b': 2},\n  e=None,\n  f: ThisObjectType,\n  """
-    expected3 = """g: I have an improper docstring,\n)"""
-
-    test_track = TestTrack()
-    actual = test_track.__repr__()
-    assert actual == expected1 + expected2 + expected3
-
-    class NoDocsTrack(utils.Track):
-        @property
-        def no_doc(self):
-            return "whee!"
-
-    bad_track = NoDocsTrack()
-    with pytest.raises(ValueError):
-        bad_track.__repr__()
